// backup 16032021 jc/src/app/modules/modulo-vacunacion-spray/models/tx-vacunacion-spray.model.ts
import { TxVacunacionSprayDetalleModel } from './tx-vacunacion-spray-detalle.model';
import { TxVacunacionSprayFotosModel } from './tx-vacunacion-spray-fotos.model';
import { TxVacunacionSprayMaquinaModel } from './tx-vacunacion-spray-maquina.model';
import { TxVacunacionSprayVacunaModel } from './tx-vacunacion-spray-vacuna.model';
import { TxVacunacionSprayResultadoModel } from './tx-vacunacion-spray-resultado.model';
export class TxVacunacionSprayModel {
    idVacunacionSpray: number;
    codigoEmpresa: string;
    descripcionEmpresa: string;
    codigoPlanta: string;
    descripcionPlanta: string;
    unidad: string;
    fecRegistro: Date;
    fecHoraRegistro: Date;
    responsableInvetsa: string;
    responsableIncubadora: string;
    flgHyLine: boolean;
    flgLohman: boolean;
    flgRoss: boolean;
    flgCobb: boolean;
    flgOtros: boolean;
    observacionOtros: string;
    flgLunes: boolean;
    flgMartes: boolean;
    flgMiercoles: boolean;
    flgJueves: boolean;
    flgViernes: boolean;
    flgSabado: boolean;
    flgDomingo: boolean;
    observacionInvetsa: string;
    observacionPlanta: string;
    promedioPollos: number;
    responsablePlanta: string;
    firmaInvetsa: string;
    firmaPlanta: string;
    flgCerrado: boolean;
    flgMigrado: boolean;
    idUsuarioCierre: number;
    fecCierre: Date;
    emailFrom: string;
    emailTo: string;
    nombreArchivo: string;
    usuarioCierre: string;
    usuarioCreacion: string;

    listarTxVacunacionSprayDetalle: TxVacunacionSprayDetalleModel[];
    listarTxVacunacionSprayFotos: TxVacunacionSprayFotosModel[];
    listarTxVacunacionSprayMaquina: TxVacunacionSprayMaquinaModel[];
    listarTxVacunacionSprayVacuna: TxVacunacionSprayVacunaModel[];
    listarTxVacunacionSprayResultado: TxVacunacionSprayResultadoModel[];

    // Auditoria
    regUsuario?: number;
    regEstacion?: string;

    constructor() {
        this.idVacunacionSpray = 0;
        this.codigoEmpresa = '';
        this.descripcionEmpresa = '';
        this.codigoPlanta = '';
        this.descripcionPlanta = '';
        this.unidad = '';
        this.fecRegistro = null;
        this.fecHoraRegistro = null;
        this.responsableInvetsa = '';
        this.responsableIncubadora = '';
        this.flgHyLine = false;
        this.flgLohman = false;
        this.flgRoss = false;
        this.flgCobb = false;
        this.flgOtros = false;
        this.observacionOtros = '';
        this.flgLunes = false;
        this.flgMartes = false;
        this.flgMiercoles = false;
        this.flgJueves = false;
        this.flgViernes = false;
        this.flgSabado = false;
        this.flgDomingo = false;
        this.observacionInvetsa = '';
        this.observacionPlanta = '';
        this.promedioPollos = 0;
        this.responsablePlanta = '';
        this.firmaInvetsa = '';
        this.firmaPlanta = '';
        this.flgCerrado = false;
        this.flgMigrado = false;
        this.idUsuarioCierre = 0;
        this.fecCierre = null;
        this.emailFrom = '';
        this.emailTo = '';
        this.nombreArchivo = '';
        this.usuarioCierre = '';
        this.usuarioCreacion = '';
    
        this.listarTxVacunacionSprayDetalle = [];
        this.listarTxVacunacionSprayFotos = [];
        this.listarTxVacunacionSprayMaquina = [];
        this.listarTxVacunacionSprayVacuna = [];
        this.listarTxVacunacionSprayResultado = [];
        this.regUsuario = 0;
        this.regEstacion = '';
    }
}